% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read "FARS" data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character sting specifying the pathname of the file located
in the local computer.}
}
\value{
This function returns a dataframe in R session after reading the
filename provided as parameter of the function
}
\description{
This is a simple function that reads csv data file using read_csv from readr
package. The function begins with checking whether the filename provided to
the function as input exists before reading it as a dataframe in r.This function
has been customized NOT to display the progress report on reading time in an
interactive session.Additionally, while creating the data in R session, this
function ignores all simple messages in the context
}
\examples{
\dontrun{
fars_read("~/data/accident_2013.csv.bz2")
}

}

