% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Check and create valid "years" of FARS data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a list or vector containing year specification such as
(\code{years=c(2013,2014,2015)}) or using (\code{years=list(2013,2014,2015)})}
}
\value{
This function returs a list of dataframes for each year as specified in
parameter (years) of the function. For example if one uses
(\code{years=c(2013,2014,2015)}) then there will be three dataframes in list
mode. Each dataframe will have two columns: Month and year.
}
\description{
This function creates multiple years of FARS data as specified by the
parameter "years." The output of the function is a list of as many dataframes
as the number of years provided in the parameter. Each dataframe contains only
two varibales, year and MONTH. Additionally, the function
also provides additional mechanism for handling errors and warnings if the
input parameter (year) is not a valid year to be found in the read files. This
error and warning mechanism is provided by using (\code{tryCatch(expr, ... )})
No messages will be posted if there is no error or wanrings.
}
\examples{
\dontrun{
fars_read_years(c(2013,2014,2015))
}

}

