% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize monthly observations of fars data by year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a list or vector containing year specification such as
(\code{years=c(2013,2014,2015)}) or using (\code{years=list(2013,2014,2015)})}
}
\value{
This function returns a dataframe, containing number of monthly observations
for each year. The monhly observations numbers are organized under columns of years,
each column representing a year.
}
\description{
This is a function to summarize multiple dataframes, each representing a year.
The summary output is returned as a dataframe, containing number of monthly
observations for each year. Given a vector of years parameter, the function
begins with reading multiple years of data using another function
(\code{fars_read_years(years)}) included in this package.Next, the function also
uses 'dplyr' package to stack dataframes (dat_list) using (\code{bind_rows(dat_list)})
and group the observations of combined dataframe by year and month using
(\code{group_by(year,MONTH)}) function, whose output is then piped for summary
dataframe using (\code{summarize(n=n())}) function of 'dplyr' package. Finally,
the output is piped to organize the monthly number of observatons by year using
(\code{spread(year,n)}) function imported from 'tidyr' package.
}
\examples{
\dontrun{
fars_summarize_years(c(2013,2014,2015))
}

}

