% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make data file name by the year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a character string or number indicating year to create the filename}
}
\value{
The function returns a character vector containing a formatted
combination of texts and parameter values \code{year}.
}
\description{
This simple function creates filename depending on the year variable that is
being provided as an input of the function. The function takes the year as
a parameter of the function. The \code{year} parameter can be supplied as just a
number or a string. Inside the function it will be converted as integer.
}
\note{
This function uses a wrapper for the C function sprintf that takes R
objects. Inside sprintf the fmt specification must include the full pathname
so that the filename returned is available in the local computer.
}
\examples{
\dontrun{
make_filename(2013)
}

}

