% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map yearly traffic accidents occured in a state}
\usage{
fars_map_state(state.num, year)
}
\description{
This function maps locations of accidents occured in each state of a specific
year. It provides messages if the state number is invalid or when there are no
accidents to plot in a state of a given year. Before creating a map for a state
all the missing observations involving Latitude and Longitude are removed from
the dataset.This function uses (\code{filter(data, state.num)}) from \code{dplyr}
package, and two other functions, one is (\code{map(state, ylim, xlim)}) from
\code{maps} package, and the other (\code{points(LONGITUD,LATITUD,pch)}) from
\code{graphics} package.
}
\details{
@param state.num A number representing a state that is converted as integer.

@param year A four-digit number representing a specific year

@importFrom dplyr filter

@importFrom maps map

@importFrom graphics points

@return The function returns a state-specific map of a year, pointing
 fatal accident locations identified by dimensions of longitude and latitude.

@examples
 \dontrun{
 fars_map_state(state.num=1,year=2015)
 }

@export
}

